# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Covid API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/reports": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "List of reports.",
                "operationId": "reports",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "The date of report in the format Y-m-d | default last added date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2020-04-16"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The query string for search by country/region and province",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "US Alabama"
                    },
                    {
                        "name": "iso",
                        "in": "query",
                        "description": "Filter by country ISO code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "USA"
                    },
                    {
                        "name": "region_name",
                        "in": "query",
                        "description": "Filter by country/region name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "US"
                    },
                    {
                        "name": "region_province",
                        "in": "query",
                        "description": "Filter by province name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Alabama"
                    },
                    {
                        "name": "city_name",
                        "in": "query",
                        "description": "Filter by city name (only for ISO code = USA)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Autauga"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTTP 200 OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/reportListPaginated"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "province": {
                "properties": {
                    "iso": {
                        "type": "string",
                        "example": "CHN"
                    },
                    "name": {
                        "type": "string",
                        "example": "Hubei"
                    },
                    "province": {
                        "type": "string",
                        "example": "Hunan"
                    },
                    "lat": {
                        "type": "string",
                        "example": "29.1832"
                    },
                    "long": {
                        "type": "string",
                        "example": "120.0934"
                    }
                },
                "type": "object"
            },
            "provinceList": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/province"
                }
            },
            "provinceListResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/provinceList"
                    }
                },
                "type": "object"
            },
...
            "reportListPaginated": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/paginator"
                    },
                    {
                        "$ref": "#/components/schemas/provinceListResponse"
                    }
                ]
            },
...
            "error422": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "HTTP Unprocessable Entity"
                    },
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "additionalProperties": {
                                "type": "string"
                            }
                        },
                        "example": {
                            "some_field": [
                                "Error description array for Some Field."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "paginator": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "/endpoint?page=3"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": "/endpoint?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "example": 21
                    },
                    "to": {
                        "type": "integer",
                        "example": 40
                    },
                    "path": {
                        "type": "string",
                        "example": "/endpoint"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "422": {
                "description": "HTTP 422 Unprocessable Entity",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error422"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "Per page for pagination.",
                "required": false,
                "schema": {
                    "type": "integer",
                    "example": 20
                }
            },
...
        }
    }
}